requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
edge-tts>=6.1.0
openai>=1.0.0
pyyaml>=6.0
mutagen>=1.47.0
